Bluetooth stuff/requirements.txt
bleak>=2.0
pyserial>=3.5
